chore: display error message in quota dialog
