+++ AdminUi/src/AdminUi/ClientApp/src/app/components/quotas/assign-quotas-dialog/assign-quotas-dialog.component.css
@@ -8,3 +8,16 @@
     justify-content: center;
     align-items: center;
 }
+
+.client-secret-warning-container {
+    margin: auto;
+    border-radius: 5px;
+    background-color: #f35f5f;
+    opacity: 0.85;
+}
+
+.client-secret-warning {
+    color: #ffffff;
+    font-weight: 500;
+    padding: 10px;
+}
