+++ AdminUi/src/AdminUi/ClientApp/src/app/components/quotas/assign-quotas-dialog/assign-quotas-dialog.component.html
@@ -31,6 +31,9 @@
         <mat-progress-spinner color="primary" mode="indeterminate"> </mat-progress-spinner>
     </div>
 </mat-dialog-content>
+<div *ngIf="errorMessage !== ''" class="client-secret-warning-container">
+    <mat-error class="client-secret-warning">{{ errorMessage }}</mat-error>
+</div>
 <mat-dialog-actions align="end">
     <button mat-flat-button mat-dialog-close>Cancel</button>
     <button mat-flat-button color="primary" (click)="assignQuota()" [disabled]="!isValid()">Confirm</button>
