+++ AdminUi/src/AdminUi/ClientApp/src/app/services/quotas-service/quotas.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from "@angular/common/http";
 import { Injectable } from "@angular/core";
-import { Observable } from "rxjs";
+import { BehaviorSubject, Observable } from "rxjs";
 import { HttpResponseEnvelope } from "src/app/utils/http-response-envelope";
 import { environment } from "src/environments/environment";
 
@@ -10,8 +10,15 @@ import { environment } from "src/environments/environment";
 export class QuotasService {
     private readonly apiUrl: string;
 
+    private readonly refreshDataSubject = new BehaviorSubject<boolean>(false);
+    public refreshData$ = this.refreshDataSubject.asObservable();
+
     public constructor(private readonly http: HttpClient) {
         this.apiUrl = environment.apiUrl;
+    }
+
+    public triggerRefresh(): void {
+        this.refreshDataSubject.next(true);
     }
 
     public getMetrics(): Observable<HttpResponseEnvelope<Metric>> {
